Refactored people and chat pane. Added button to copy invite link in mobile people pane.
