+++ packages/react-composites/src/composites/CallWithChatComposite/styles/EmbeddedPeoplePane.styles.ts
@@ -1,0 +1,18 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT license.
+
+import { IButtonStyles } from '@fluentui/react';
+
+export const copyLinkButtonStyles: IButtonStyles = {
+  root: {
+    height: '3rem',
+    border: 'none',
+    borderRadius: '0.5rem',
+    margin: '0.75rem 1rem'
+  },
+  textContainer: {
+    display: 'contents'
+  }
+};
+
+export const linkIconStyles = { marginRight: '0.5rem' };
